+++ src/utils/index.ts
@@ -6,14 +6,19 @@ export * from './formatData/stringOperation';
 export * from './formatData/dateOperations';
 export * from './tools/Notification';
 export * from './info/useBitcoinRates';
-export * from './tools/helperFunctions';
+export * from './formatData/helperFunctions';
 export * from './formatData/sha256';
 export * from './info/getRandom';
+
 export * from './validations/need';
 export * from './validations/hasShipping';
 export * from './validations/hasOnChain';
-export * from './lnUrl';
+export * from './validations/project';
 export * from './validations/hasTwitterAccount';
 export * from './validations/testImage';
 export * from './validations/regex';
+
+export * from './lnUrl';
 export * from './formatData/aggregateTransactions';
+export * from './unitConversion/typeConversion';
+export * from './unitConversion/BTCConversionUtils';
