+++ src/utils/unitConversion/typeConversion.ts
@@ -1,0 +1,3 @@
+export const toInt = (val: any) => {
+  return parseInt(`${val}`, 10);
+};
