+++ src/utils/validations/project.ts
@@ -1,0 +1,5 @@
+import { Project, ProjectStatus } from '../../types/generated/graphql';
+
+export const isProjectActive = (project: Project) => {
+  return project.status === ProjectStatus.Active;
+};
